requirements.txt
pytest==7.4.0
selenium==4.12.0
allure-pytest==2.24.0

